requirements.txt
absl-py==2.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
bcrypt==4.3.0
black==23.3.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
DateTime==5.1
Django==5.0.6
django-ckeditor==6.6.1
django-cors-headers==4.6.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-js-asset==2.1.0
django-safedelete==1.3.2
django-session-timeout==0.1.0
django-solo==2.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-extensions==0.7.1
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
drf-yasg==1.21.8
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==2.0.2
Faker==28.0.0
fastapi==0.115.4
fastapi-cli==0.0.5
fastapi-utils==0.8.0
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.10.0
gast==0.6.0
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.69.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keras==3.8.0
keybert==0.8.5
libclang==18.1.1
linkify-it-py==2.0.3
Mako==1.3.8
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
ml-dtypes==0.4.1
mock==5.1.0
model-bakery==1.11.0
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
nltk==3.9.1
numpy==2.0.2
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.10
packaging==23.1
pandas==2.2.3
passlib==1.7.4
path==16.7.1
path.py==12.5.0
pathspec==0.11.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.0.0
proto-plus==1.24.0
protobuf==5.28.0
psutil==5.9.8
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.24.12
pyOpenSSL==23.2.0
pyparsing==3.1.4
pytest==7.3.1
pytest-django==4.5.2
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.16
pytz==2023.3
PyYAML==6.0
referencing==0.30.0
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.30
ruamel.yaml.clib==0.2.7
ruff==0.6.2
SafeDelete==0.0.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
shellingham==1.5.4
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.4.4
starlette==0.41.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.3.0
textual==0.83.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.0.1
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.3
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==2.3.6
wrapt==1.17.1
zipp==3.15.0
zope.interface==6.0
